move and tweak bloomery class
